Refactor add-shares function to separate parser logic from adding shares to DB
